+++ server/controllers/userController.js
@@ -3,7 +3,8 @@ import { getQueryData } from "../utils/getQueryData.js";
 import addLinks from "../utils/addLinks.js";
 import productModel from "../models/productModel.js";
 import articleModel from "../models/articleModel.js";
-import { checkBodyErrors } from "../utils/checkBodyErrors.js";
+import { validationResult } from "express-validator";
+import { articleService } from "../services/articleService.js";
 
 export const getUsers = async (req, res, next) => {
   try {
@@ -37,6 +38,15 @@ export const getMe = async (req, res, next) => {
 
 export const saveProduct = async (req, res, next) => {
   try {
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        status: "fail",
+        errors: errors.array(),
+      });
+    }
+
     let data = await userService.addToArray(
       req.user._id,
       "savedProducts",
@@ -55,6 +65,15 @@ export const saveProduct = async (req, res, next) => {
 };
 export const discardProduct = async (req, res, next) => {
   try {
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        status: "fail",
+        errors: errors.array(),
+      });
+    }
+
     let data = await userService.deleteFromArray(
       req.user._id,
       "savedProducts",
@@ -73,12 +92,22 @@ export const discardProduct = async (req, res, next) => {
 
 export const likeArticle = async (req, res, next) => {
   try {
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        status: "fail",
+        errors: errors.array(),
+      });
+    }
     let data = await userService.addToArray(
       req.user._id,
       "likedArticles",
       req.body.id,
       articleModel
     );
+
+    await articleService.updateOne(req.body.id, { $inc: { likes: 1 } });
 
     data = addLinks(req, data[0], "avatar");
     res.status(200).json({
@@ -91,12 +120,22 @@ export const likeArticle = async (req, res, next) => {
 };
 export const unlikeArticle = async (req, res, next) => {
   try {
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        status: "fail",
+        errors: errors.array(),
+      });
+    }
     let data = await userService.deleteFromArray(
       req.user._id,
       "likedArticles",
       req.params.id
     );
     data = addLinks(req, data[0], "avatar");
+
+    await articleService.updateOne(req.params.id, { $inc: { likes: -1 } });
 
     res.status(200).json({
       status: "success",
@@ -109,12 +148,21 @@ export const unlikeArticle = async (req, res, next) => {
 
 export const saveArticle = async (req, res, next) => {
   try {
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        status: "fail",
+        errors: errors.array(),
+      });
+    }
     let data = await userService.addToArray(
       req.user._id,
       "savedArticles",
       req.body.id,
       articleModel
     );
+
     data = addLinks(req, data[0], "avatar");
 
     res.status(200).json({
@@ -127,6 +175,14 @@ export const saveArticle = async (req, res, next) => {
 };
 export const discardArticle = async (req, res, next) => {
   try {
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        status: "fail",
+        errors: errors.array(),
+      });
+    }
     let data = await userService.deleteFromArray(
       req.user._id,
       "savedArticles",
@@ -145,14 +201,17 @@ export const discardArticle = async (req, res, next) => {
 
 export const addToCart = async (req, res, next) => {
   try {
-    const prodData = { id: req.body.id, quantity: req.body.quantity };
+    const errors = validationResult(req);
 
-    let data = await userService.addToArray(
-      req.user._id,
-      "cart",
-      prodData,
-      productModel
-    );
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        status: "fail",
+        errors: errors.array(),
+      });
+    }
+    const { id, quantity } = req.body;
+
+    let data = await userService.addToCart(req.user._id, id, quantity);
     data = addLinks(req, data[0], "avatar");
 
     res.status(200).json({
@@ -166,7 +225,15 @@ export const addToCart = async (req, res, next) => {
 
 export const deleteFromCart = async (req, res, next) => {
   try {
-    let data = await userService.deleteFromArray(req.user._id, "cart", req.params.id);
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        status: "fail",
+        errors: errors.array(),
+      });
+    }
+    let data = await userService.deleteFromCart(req.user._id, req.params.id);
     data = addLinks(req, data[0], "avatar");
     res.status(200).json({
       status: "success",
@@ -179,6 +246,14 @@ export const deleteFromCart = async (req, res, next) => {
 
 export const updateCartItem = async (req, res, next) => {
   try {
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        status: "fail",
+        errors: errors.array(),
+      });
+    }
     let data = await userService.updateCart(
       req.user._id,
       req.params.id,
@@ -197,6 +272,14 @@ export const updateCartItem = async (req, res, next) => {
 
 export const updateUserInfo = async (req, res, next) => {
   try {
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        status: "fail",
+        errors: errors.array(),
+      });
+    }
     const userInfo = {
       name: req.body?.name,
       lastName: req.body?.lastName,
@@ -226,7 +309,14 @@ export const deleteUser = async (req, res, next) => {
 
 export const changeRole = async (req, res, next) => {
   try {
-    checkBodyErrors(req, res);
+    const errors = validationResult(req);
+
+    if (!errors.isEmpty()) {
+      return res.status(400).json({
+        status: "fail",
+        errors: errors.array(),
+      });
+    }
 
     let data = await userService.updateOne(req.params.id, { role: req.body.role });
